requirements.txt
boto3==1.12.30
botocore==1.15.30
CacheControl==0.12.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
firebase-admin==3.2.1
fuzzywuzzy==0.18.0
gensim==3.8.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
httplib2==0.17.0
idna==2.9
jmespath==0.9.5
kiwisolver==1.1.0
matplotlib==3.2.1
msgpack==1.0.0
networkx==2.4
nltk==3.4.5
numpy==1.18.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
scipy==1.4.1
six==1.14.0
smart-open==1.10.0
uritemplate==3.0.1
urllib3==1.25.8
